requirements.txt
mysqlclient
pillow
django-import-export
django-simpleui
django-nested-admin
setuptools
django-adminlte3
